+++ crates/builder/src/database/mod.rs
@@ -1,4 +1,5 @@
 pub mod bal_builder_db;
 pub mod bundle_db;
+pub mod shared_state_provider_db;
 pub mod temporal_db;
 pub mod temporal_map;
